Bugfix added to get_flows
Bug addressed: the cartesian product of origin, destination, was resulting in od-pairs that were not observed and therefore were not necessary to be included in the final flow dataset. Instead, we want the cartesian product of unique od-pairs that are observed in the data. This brings down the amount of data stored by an order of magnitude.